Fast_API/requirements.txt
fastapi
spacy
pydantic
uvicorn
python-dotenv
gunicorn
pinecone-client==2.2.4
python-jose
passlib
asyncpg
snowflake-connector-python
pandas
python-multipart
